Implemented toasts, also reorganization of Upload
