+++ frontend/src/App.js
@@ -2,11 +2,12 @@ import React, { Component } from "react";
 import { BrowserRouter, Link, Switch } from "react-router-dom";
 import { Route } from "react-router";
 import {
-    DigitDialog,
-    DigitProviders,
-    DigitHeader,
-    DigitNavLink,
-    DigitText
+  DigitDialog,
+  DigitProviders,
+  DigitHeader,
+  DigitNavLink,
+  DigitToast,
+  DigitText
 } from "@cthit/react-digit-components";
 
 import Router from "./components/Router";
@@ -14,35 +15,33 @@ import NotFound from "./components/NotFound";
 import "./App.css";
 
 class App extends Component {
-    render() {
-        return (
-            <DigitProviders
-                preloadedState={{
-                    loading: true
-                }}
-                defaultLanguage="en"
-            >
-                <DigitHeader
-                    title="TasteIT"
-                    renderDrawer={closeDrawer => (
-                        <div className="headerNavLinks">
-                            <DigitNavLink
-                                text="View Recipe"
-                                link="/"
-                                onClick={closeDrawer}
-                            />
-                            <DigitNavLink
-                                text="Upload recipe"
-                                link="/upload"
-                                onClick={closeDrawer}
-                            />
-                        </div>
-                    )}
-                    renderMain={() => <Router />}
+  render() {
+    return (
+      <DigitProviders>
+        <React.Fragment>
+          <DigitToast />
+          <DigitHeader
+            title="TasteIT"
+            renderDrawer={closeDrawer => (
+              <div className="headerNavLinks">
+                <DigitNavLink
+                  text="View Recipe"
+                  link="/"
+                  onClick={closeDrawer}
                 />
-            </DigitProviders>
-        );
-    }
+                <DigitNavLink
+                  text="Upload recipe"
+                  link="/upload"
+                  onClick={closeDrawer}
+                />
+              </div>
+            )}
+            renderMain={() => <Router />}
+          />
+        </React.Fragment>
+      </DigitProviders>
+    );
+  }
 }
 
 export default App;
